requirements.txt
Pillow
python-barcode
pystrich
Flask
flask-cors
gunicorn